+++ openCVLibrary343/src/main/java/org/opencv/calib3d/Calib3d.java
@@ -3,8 +3,6 @@
 //
 package org.opencv.calib3d;
 
-import java.util.ArrayList;
-import java.util.List;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfDouble;
 import org.opencv.core.MatOfPoint2f;
@@ -14,6 +12,9 @@ import org.opencv.core.Rect;
 import org.opencv.core.Size;
 import org.opencv.core.TermCriteria;
 import org.opencv.utils.Converters;
+
+import java.util.ArrayList;
+import java.util.List;
 
 // C++: class Calib3d
 //javadoc: Calib3d
@@ -1285,11 +1286,11 @@ public class Calib3d {
         double[] principalPoint_out = new double[2];
         double[] aspectRatio_out = new double[1];
         calibrationMatrixValues_0(cameraMatrix.nativeObj, imageSize.width, imageSize.height, apertureWidth, apertureHeight, fovx_out, fovy_out, focalLength_out, principalPoint_out, aspectRatio_out);
-        if(fovx!=null) fovx[0] = (double)fovx_out[0];
-        if(fovy!=null) fovy[0] = (double)fovy_out[0];
-        if(focalLength!=null) focalLength[0] = (double)focalLength_out[0];
+        if(fovx!=null) fovx[0] = fovx_out[0];
+        if(fovy!=null) fovy[0] = fovy_out[0];
+        if(focalLength!=null) focalLength[0] = focalLength_out[0];
         if(principalPoint!=null){ principalPoint.x = principalPoint_out[0]; principalPoint.y = principalPoint_out[1]; } 
-        if(aspectRatio!=null) aspectRatio[0] = (double)aspectRatio_out[0];
+        if(aspectRatio!=null) aspectRatio[0] = aspectRatio_out[0];
         return;
     }
