+++ openCVLibrary343/src/main/java/org/opencv/imgproc/Imgproc.java
@@ -3,9 +3,6 @@
 //
 package org.opencv.imgproc;
 
-import java.lang.String;
-import java.util.ArrayList;
-import java.util.List;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfFloat;
 import org.opencv.core.MatOfInt;
@@ -18,9 +15,10 @@ import org.opencv.core.RotatedRect;
 import org.opencv.core.Scalar;
 import org.opencv.core.Size;
 import org.opencv.core.TermCriteria;
-import org.opencv.imgproc.CLAHE;
-import org.opencv.imgproc.LineSegmentDetector;
 import org.opencv.utils.Converters;
+
+import java.util.ArrayList;
+import java.util.List;
 
 // C++: class Imgproc
 //javadoc: Imgproc
@@ -636,7 +634,7 @@ public class Imgproc {
     {
         double[] response_out = new double[1];
         Point retVal = new Point(phaseCorrelate_0(src1.nativeObj, src2.nativeObj, window.nativeObj, response_out));
-        if(response!=null) response[0] = (double)response_out[0];
+        if(response!=null) response[0] = response_out[0];
         return retVal;
     }
